# src/backend_go/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Cognix Golang API Documentation",
        "title": "Cognix API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/config-map/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "returns list of values from given config map name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigMap"
                ],
                "summary": "returns list of values from given config map name",
                "operationId": "configmap_get_config_map",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of config map",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/proto.ConfigMapRecord"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create or update value in config map",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigMap"
                ],
                "summary": "create or update value in config map",
                "operationId": "configmap_save_config_map",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of config map",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "payload",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.ConfigMapValue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.JsonResponse"
                        }
                    }
                }
            }
        },
        "/api/config-map/{name}/{key}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "deletes value from config map",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigMap"
                ],
                "summary": "deletes value from config map",
                "operationId": "configmap_delete_config_map",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of config map",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "key of config map",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.JsonResponse"
                        }
                    }
                }
            }
        },
...
        "/manage/connector": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of allowed connectors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "return list of allowed connectors",
                "operationId": "connectors_get_all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Connector"
                            }
                        }
                    }
                }
            }
        },
        "/manage/connector/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "creates connector",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "creates connector",
                "operationId": "connectors_create",
                "parameters": [
                    {
                        "description": "connector create parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.CreateConnectorParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Connector"
                        }
                    }
                }
            }
        },
        "/manage/connector/source_types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of source types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "return list of source types",
                "operationId": "connectors_get_source_types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SourceTypeDescription"
                            }
                        }
                    }
                }
            }
        },
        "/manage/connector/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of allowed connectors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "return list of allowed connectors",
                "operationId": "connectors_get_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "connector id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Connector"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "updates connector",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "updates connector",
                "operationId": "connectors_update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "connector id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "connector update parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.UpdateConnectorParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Connector"
                        }
                    }
                }
            }
        },
        "/manage/connector/{id}/{action}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete or restore connector",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connectors"
                ],
                "summary": "delete or restore connector",
                "operationId": "Connectors_delete_restore",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Connectors id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "action : delete | restore ",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Connector"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "decimal.NullDecimal": {
            "type": "object",
            "properties": {
                "decimal": {
                    "type": "number"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Connector": {
            "type": "object",
            "properties": {
                "connector_specific_config": {
                    "$ref": "#/definitions/model.JSONMap"
                },
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "docs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Document"
                    }
                },
                "docs_map": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.Document"
                    }
                },
                "id": {
                    "type": "number"
                },
                "last_successful_analysis": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "name": {
                    "type": "string"
                },
                "refresh_freq": {
                    "type": "integer"
                },
                "source": {
                    "$ref": "#/definitions/model.SourceType"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "total_docs_indexed": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.Document": {
            "type": "object",
            "properties": {
                "analyzed": {
                    "type": "boolean"
                },
                "chunking_session": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "connector_id": {
                    "type": "number"
                },
                "creation_date": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "original_url": {
                    "type": "string"
                },
                "parent_id": {
                    "$ref": "#/definitions/decimal.NullDecimal"
                },
                "signature": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
...
        "model.SourceType": {
            "type": "string",
            "enum": [
                "ingestion_api",
                "slack",
                "web",
                "google_drive",
                "gmail",
                "requesttracker",
                "github",
                "gitlab",
                "guru",
                "bookstack",
                "confluence",
                "slab",
                "jira",
                "productboard",
                "file",
                "notion",
                "zulip",
                "linear",
                "hubspot",
                "document360",
                "gong",
                "google_sites",
                "zendesk",
                "loopio",
                "sharepoint",
                "one-drive",
                "msteams",
                "youtube"
            ],
            "x-enum-varnames": [
                "SourceTypeIngestionApi",
                "SourceTypeSlack",
                "SourceTypeWEB",
                "SourceTypeGoogleDrive",
                "SourceTypeGMAIL",
                "SourceTypeRequesttracker",
                "SourceTypeGithub",
                "SourceTypeGitlab",
                "SourceTypeGuru",
                "SourceTypeBookstack",
                "SourceTypeConfluence",
                "SourceTypeSlab",
                "SourceTypeJira",
                "SourceTypeProductboard",
                "SourceTypeFile",
                "SourceTypeNotion",
                "SourceTypeZulip",
                "SourceTypeLinear",
                "SourceTypeHubspot",
                "SourceTypeDocument360",
                "SourceTypeGong",
                "SourceTypeGoogleSites",
                "SourceTypeZendesk",
                "SourceTypeLoopio",
                "SourceTypeSharepoint",
                "SourceTypeOneDrive",
                "SourceTypeMsTeams",
                "SourceTypeYoutube"
            ]
        },
        "model.SourceTypeDescription": {
            "type": "object",
            "properties": {
                "id": {
                    "$ref": "#/definitions/model.SourceType"
                },
                "isImplemented": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "parameters.ConfigMapValue": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "parameters.CreateConnectorParam": {
            "type": "object",
            "properties": {
                "connector_specific_config": {
                    "$ref": "#/definitions/model.JSONMap"
                },
                "disabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "refresh_freq": {
                    "type": "integer"
                },
                "shared": {
                    "type": "boolean"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "parameters.UpdateConnectorParam": {
            "type": "object",
            "properties": {
                "connector_specific_config": {
                    "$ref": "#/definitions/model.JSONMap"
                },
                "name": {
                    "type": "string"
                },
                "refresh_freq": {
                    "type": "integer"
                },
                "shared": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "pg.NullTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "proto.ConfigMapRecord": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "server.JsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}